added test cases for row and column
